requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aioitertools==0.12.0
aiosignal==1.3.2
aiosqlite==0.20.0
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arize-phoenix==7.9.0
arize-phoenix-evals==0.19.0
arize-phoenix-otel==0.6.1
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==24.3.0
Authlib==1.4.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
comm==0.2.2
cryptography==43.0.3
dataclasses-json==0.6.7
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.15
dirtyjson==1.0.8
distro==1.9.0
docx2txt==0.8
executing==2.1.0
fastapi==0.115.6
fastjsonschema==2.21.1
filelock==3.17.0
filetype==1.2.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.12.0
funcy==2.0
future==1.0.0
gkeepapi==0.15.1
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
gpsoauth==1.1.1
graphql-core==3.2.5
greenlet==3.1.1
grpc-interceptor==0.15.4
grpcio==1.69.0
gspread==6.1.4
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.27.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
litellm==1.59.3
llama-index-callbacks-arize-phoenix==0.3.0
llama-index-core==0.12.11
llama-index-embeddings-ollama==0.5.0
llama-index-llms-lmstudio==0.3.0
llama-index-llms-ollama==0.5.0
llama-index-readers-file==0.4.3
llama-index-readers-google==0.5.0
llama-index-storage-docstore-redis==0.3.0
llama-index-storage-index-store-redis==0.4.0
llama-index-storage-kvstore-redis==0.3.0
llama-index-vector-stores-redis==0.4.0
Mako==1.3.8
MarkupSafe==3.0.2
marshmallow==3.25.1
matplotlib-inline==0.1.7
mistune==3.1.0
ml-dtypes==0.4.1
multidict==6.1.0
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.3.2
notebook_shim==0.2.4
numpy==2.2.1
oauth2client==4.1.3
oauthlib==3.2.2
ollama==0.4.6
openai==1.59.9
openinference-instrumentation==0.1.20
openinference-instrumentation-llama-index==3.1.3
openinference-semantic-conventions==0.1.12
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.10.5
pydantic_core==2.27.2
PyDrive==1.3.1
PyDrive2==1.21.3
Pygments==2.19.1
PyMuPDF==1.25.2
pyOpenSSL==24.2.1
pyparsing==3.2.1
pypdf==5.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
redis==5.2.1
redisvl==0.3.8
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.22.3
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.1
Send2Trash==1.8.3
setuptools==75.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
sqlean.py==3.47.0
stack-data==0.6.3
starlette==0.41.3
strawberry-graphql==0.253.1
striprtf==0.0.26
tabulate==0.9.0
tenacity==9.0.0
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.4.0
tokenizers==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.34.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
wheel==0.44.0
widgetsnbextension==4.0.13
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
